requirements.txt
applescript==2021.2.9
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyYAML==6.0
validators==0.24.0
